requirements.txt
streamlit
moviepy
pydub
google-cloud-speech
google-cloud-texttospeech
openai
python-dotenv
